requirements.txt
pandas
scikit-learn
numpy
Pillow
joblib
opencv-python
pyarrow
